[Commit-3] Tweaked react homepage and restructured some folders
